requirements.txt
discord.py==2.6.4
dotenv==1.2.1